Замена int на long

+++ src/Main.java
@@ -2,10 +2,8 @@ public class Main {
     public static void main(String[] args) {
 
 
-        SalesManager oSale = new SalesManager(new int []{});
+        SalesManager oSale = new SalesManager(new long[]{});
         oSale.max();
-
-
 
 
     }
